// src/components/pages/Users/User/types.ts
export enum Label {
  EDIT = "Edit",
  FETCHING_USER = "Fetching user data...",
  FETCHING_USER_ERROR = "Couldn't fetch user data.",
  NO_USER = "User not found.",
  TAB_ACCOUNT_MANAGEMENT = "Account management",
  TAB_ENTITLEMENTS = "Entitlements",
  TAB_GROUPS = "Groups",
  TAB_ROLES = "Roles",
  TAB_SUMMARY = "Summary",
}